api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
bleach==6.0.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docutils==0.19
fake-useragent==1.1.1
Flask==2.3.2
fonttools==4.39.3
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
grequests==0.6.0
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
keyring==23.13.1
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pkginfo==1.9.6
progress==1.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.1
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
SimpleWebSocketServer==0.1.2
six==1.16.0
sklearn==0.0.post4
threadpoolctl==3.1.0
tomli==2.0.1
twine==4.0.2
typing==3.7.4.3
tzdata==2023.3
urllib3==1.26.15
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
